pyproject.toml
[tool.poetry]
name = "stock-bot"
version = "0.0.1"
description = ""
authors = ["Andrea Scorza"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.1"
python-telegram-bot = "^21.6"
sqlalchemy = "^2.0.35"

[tool.poetry.scripts]
start-bot = "stock_bot.bot:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
